trying MVVM for paint screen.
